refactor: 4계층 구조 수정
-service 이동 (application)
