+++ src/main/java/org/cheonyakplanet/be/application/dto/infra/InfraResponseDTO.java
@@ -1,0 +1,18 @@
+package org.cheonyakplanet.be.application.dto.infra;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+
+public class InfraResponseDTO {
+	private List<StationDTO> stations;
+	private List<SchoolDTO> schools;
+}
