+++ src/main/java/org/cheonyakplanet/be/presentation/controller/HomeController.java
@@ -3,9 +3,9 @@ package org.cheonyakplanet.be.presentation.controller;
 import java.util.List;
 
 import org.cheonyakplanet.be.application.dto.ApiResponse;
+import org.cheonyakplanet.be.application.service.CommunityService;
+import org.cheonyakplanet.be.application.service.SubscriptionService;
 import org.cheonyakplanet.be.domain.entity.Post;
-import org.cheonyakplanet.be.domain.service.CommunityService;
-import org.cheonyakplanet.be.domain.service.SubscriptionService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
